+++ src/rex.forms/src/rex/forms/interface/__init__.py
@@ -5,10 +5,12 @@
 
 from .draftform import *
 from .form import *
+from .presentation_adaptor import *
 
 
 __all__ = (
     'DraftForm',
     'Form',
+    'PresentationAdaptor',
 )